Added half sibling recipe and more improvements to threshold analysis

+++ src/main/java/uk/ac/standrews/cs/population_linkage/aleks/PredicateEfficacy.java
@@ -57,7 +57,7 @@ public class PredicateEfficacy {
                 long trueCount = trueMatches.get(0);
                 long totalCount = total.get(0);
 
-                System.out.println(s + " " + trueCount + "/" + totalCount + " " + trueCount / totalCount);
+                System.out.println(s + " " + trueCount + "/" + totalCount + " " + (double) (trueCount / totalCount));
             }
         }
     }
